chore: 주석 및 불필요한 코드 삭제
