--- client/src/constants/signUp.ts
@@ -1,7 +1,0 @@
-export const SIGN_UP_MESSAGES = {
-  LOGIN_ID_GUIDE: '다른 유저와 겹치지 않도록 입력해주세요. (4~10자)',
-  PASSWORD_GUIDE:
-    '영문 소문자와 숫자, 특수기호(!@#$%^&*())를 모두 포함하여 입력해주세요. (8~16자)',
-  MUST_FILL_THE_BOX: '필수 입력 항목입니다.',
-  POLICY_GUIDE: '선택항목에 대한 동의 포함',
-};
