--- client/src/constants/updateUser.ts
@@ -1,5 +1,0 @@
-export const USER_UPDATE_MESSAGES = {
-  NICKNAME_GUIDE: '닉네임을 수정할 경우에 입력해주세요.',
-  PASSWORD_GUIDE: '비밀번호를 수정할 경우에 입력해주세요.',
-  PASSWORD_CONFIRM_GUIDE: '비밀번호를 다시 입력해주세요.',
-};
